DivNoise/requirements.txt
Flask==1.1.4
numpy==1.14.5
Pillow==4.2.1
scikit-image==0.13.0
scikit-learn==0.19.1
scipy==0.19.1
protobuf==3.19.6
tensorflow==1.2.1
#opencv-python==4.4.0.42